refactor: API 통신 책임 / 도메인 처리 책임 분리
