+++ src/main/java/roomescape/reservation/dto/PaymentApiResponse.java
@@ -2,9 +2,9 @@ package roomescape.reservation.dto;
 
 import roomescape.reservation.model.Payment;
 
-public record PaymentApiResponse(String status, String paymentKey, String orderId) {
+public record PaymentApiResponse(String status, String orderId, String paymentKey) {
 
     public Payment toEntity(Long amount) {
-        return new Payment(status, paymentKey, orderId, amount);
+        return new Payment(status, orderId, paymentKey, amount);
     }
 }
